+++ bundle_windows.go
@@ -1,0 +1,5 @@
+package hashcatlauncher
+
+func (a *App) Bundle() error {
+	return nil
+}