+++ server.go
@@ -2,7 +2,6 @@ package hashcatlauncher
 
 import (
 	"embed"
-	"log"
 	"net"
 	"net/http"
 )
@@ -10,13 +9,15 @@ import (
 //go:embed frontend/hashcat.launcher/build
 var fs embed.FS
 
-func (a *App) NewServer() {
+func (a *App) NewServer() error {
 	ln, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	go http.Serve(ln, http.FileServer(http.FS(fs)))
 
 	a.Server = ln
+
+	return nil
 }